+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Fixed
+- partiql-types: Fixed handling of struct fields to be resilient to field order w.r.t. equality and hashing 
 
 ## [0.7.1] - 2024-03-15
 ### Changed